syncoldgetdatasetfunc
